Add raw in-memory streams for compress/decompress
This commit exposes the raw `mz_stream` type as `Compress` or `Decompress` which
are thin wrappers around the raw operations on these sorts of streams. This
should provide the ability to have a much finer grain of control over how
compression is done or how decompression is done (especially around buffer
management).

Closes #14