Add a `bufread` module for buffered types
Right now all stream types in `read` have an unconditional buffer between the
underlying stream and what we're passing down to the compression layer. This
buffer, however, is redundant for types that implement the `BufRead` trait.

This commit adds explicit support for `BufRead` types which have their own local
buffer, and then all other types are built on top of these types. A local
implementation of `BufReader` is vendored for the `reset()` method.