Implemented multi member gz decoder
Added a MultiDecoderReader and MultiDecoderReaderBuf to gz. I moved
the read header code to a separate function to avoid code duplication.

In order to keep the reader but reset the decoder a reset_data method
has been added to deflate.rs and a reset method to the CrcReader.

Tests for multimember gz files are in the test directory.