Expose the underlying mz_stream from Compress and Decompress.
zlib exposes more API surface than miniz, and I needed to use a few of those APIs (like inflateReset),
but using the libz-sys crate directly means I have to reimplement a number of things from flate2.

Really all I need is to be able to get the underlying mz_stream so I can call the zlib APIs on it.
This patch adds a `get_raw` method to Compress and Decompress that returns a `*mut mz_state` which can
be passed directly to the functions defined in libz-sys. This doesn't add much complexity to flate2, but
fixes my niche use case.