Mention rust backend & update miniz and flate info
Note that there is now a rust back-end.

Point the miniz link to the new github page instead of the removed google code one.

Mention that the flate library is no longer part of the rust compiler, and fix the link so it doesn't point to the wrong library.
(the libflate library on crates.io is a different library.)