Emit a RFC 1952 compatible XFL flag again
RFC 1952 specifies the following values
for the XFL flag:

* 2: maximum compression, slowest
* 4: fastest algorithm config

Commit 7e0390b51336b9737eefb10d1afc14d5e311fb9d has
changed our behaviour from conforming to that
to setting the XFL flag to 9 when compression
is best, and 0 when it is the fastest one.

We revert that change in order to have consistent
output with pre 1.0.0 versions of flate2 and also
to be consistent with RFC 1952 and the ecosystem
around it, e.g. unix tooling that assumes the
RFC specified layout for the XFL flag.

Fixes #137.