Configure allocation/free functions in zlib/miniz
This commit configures the allocation/free functions of zlib/miniz to
run through the same allocator as the rest of Rust's memory, ensuring
that we're all using the same global Rust allocator for everything zlib
needs.

This also fixes an issue where the initialization of `mz_stream` was
undefined behavior with zlib activated because it created null function
pointers.