Update to libz-sys 1.1.0, and support use with zlib-ng (#248)
libz-sys 1.1.0 introduces optional support for zlib-ng, if the default
`stock-zlib` feature is disabled and the `zlib-ng` feature is enabled.
Modify the dependency on libz-sys to disable the `stock-zlib` feature,
which allows crates using flate2 to opt into zlib-ng if they wish. Add a
`zlib-ng-compat` feature to flate2, to opt into zlib-ng explicitly.

Update the README to document this, and to describe some of the
tradeoffs between zlib implementations.