Introduce a changelog with `cargo changelog`
This is a proposal for maintaining a changelog, which currently does not exist.

The way this works is to use conventional commits *only* when those notes should
appear in the changelog next time it is updated with `cargo changelog` or `cargo smart-release`.

This way, useful changelog can be generated with minimal effort, while making it possible to
hand-write portions of it as one sees fit.