bump *tested* MSRV to 1.63
That way it should compile with minimal versions.

Reason for this is the following:

````
error: package `cc v1.0.98` cannot be built because it requires rustc 1.63 or newer, while the currently active rustc version is 1.56.1
````

Note that the expressed MSRV is still a lower one as it may work for
some people that use different settings and/or configuration.